# docs/api/openapi-live.json
{"openapi":"3.1.0","info":{"title":"ScanDetection","version":"0.1.0"},"paths":{"/api/v1/health":{"get":{"tags":["health"],"summary":"Health Check","operationId":"health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Health Check Api V1 Health Get"}}}}}}},"/api/v1/verify":{"post":{"tags":["verify"],"summary":"Verify","operationId":"verify_api_v1_verify_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_verify_api_v1_verify_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/preprocess":{"post":{"tags":["preprocess"],"summary":"Preprocess","operationId":"preprocess_api_v1_preprocess_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_preprocess_api_v1_preprocess_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PreprocessResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/detect":{"post":{"tags":["detect"],"summary":"Detect","operationId":"detect_api_v1_detect_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_detect_api_v1_detect_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DetectResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/judge":{"post":{"tags":["judge"],"summary":"Judge","operationId":"judge_api_v1_judge_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JudgeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JudgeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/batch":{"post":{"tags":["batch"],"summary":"Submit Batch","operationId":"submit_batch_api_v1_batch_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/batch/{batch_id}":{"get":{"tags":["batch"],"summary":"Batch Status","operationId":"batch_status_api_v1_batch__batch_id__get","parameters":[{"name":"batch_id","in":"path","required":true,"schema":{"type":"string","title":"Batch Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/batch/{batch_id}/cancel":{"post":{"tags":["batch"],"summary":"Cancel Batch","operationId":"cancel_batch_api_v1_batch__batch_id__cancel_post","parameters":[{"name":"batch_id","in":"path","required":true,"schema":{"type":"string","title":"Batch Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/review":{"post":{"tags":["review"],"summary":"Review","operationId":"review_api_v1_review_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/explain":{"post":{"tags":["explain"],"summary":"Explain","operationId":"explain_api_v1_explain_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/report":{"post":{"tags":["report"],"summary":"Report","operationId":"report_api_v1_report_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_report_api_v1_report_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/report/{report_id}/pdf":{"get":{"tags":["report"],"summary":"Report Pdf","operationId":"report_pdf_api_v1_report__report_id__pdf_get","parameters":[{"name":"report_id","in":"path","required":true,"schema":{"type":"string","title":"Report Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/records":{"get":{"tags":["records"],"summary":"Records","operationId":"records_api_v1_records_get","parameters":[{"name":"level","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Level"}},{"name":"class","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Class"}},{"name":"from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"From"}},{"name":"to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"To"}},{"name":"workpiece","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Workpiece"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Size"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecordsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/models":{"get":{"tags":["models"],"summary":"List Models","operationId":"list_models_api_v1_models_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/models/{model_id}/activate":{"post":{"tags":["models"],"summary":"Activate Model","operationId":"activate_model_api_v1_models__model_id__activate_post","parameters":[{"name":"model_id","in":"path","required":true,"schema":{"type":"string","title":"Model Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/audit":{"get":{"tags":["audit"],"summary":"Audit","operationId":"audit_api_v1_audit_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_detect_api_v1_detect_post":{"properties":{"image":{"type":"string","contentMediaType":"application/octet-stream","title":"Image"},"pixel_spacing_mm":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Pixel Spacing Mm"},"conf":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Conf"}},"type":"object","required":["image"],"title":"Body_detect_api_v1_detect_post"},"Body_preprocess_api_v1_preprocess_post":{"properties":{"image":{"type":"string","contentMediaType":"application/octet-stream","title":"Image"},"gamma":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Gamma"}},"type":"object","required":["image"],"title":"Body_preprocess_api_v1_preprocess_post"},"Body_report_api_v1_report_post":{"properties":{"image":{"anyOf":[{"type":"string","contentMediaType":"application/octet-stream"},{"type":"null"}],"title":"Image"},"image_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Id"},"pixel_spacing_mm":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Pixel Spacing Mm"},"base_metal_thickness_mm":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Base Metal Thickness Mm"},"standard_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Standard Id"},"iqi_roi":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Iqi Roi"},"workpiece_no":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Workpiece No"},"weld_no":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Weld No"},"signer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Signer"},"template":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Template"}},"type":"object","title":"Body_report_api_v1_report_post"},"Body_verify_api_v1_verify_post":{"properties":{"image":{"type":"string","contentMediaType":"application/octet-stream","title":"Image"},"iqi_roi":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Iqi Roi"}},"type":"object","required":["image"],"title":"Body_verify_api_v1_verify_post"},"DefectOut":{"properties":{"id":{"type":"string","title":"Id"},"class_id":{"type":"integer","title":"Class Id"},"class_name":{"type":"string","title":"Class Name"},"bbox":{"items":{"type":"number"},"type":"array","title":"Bbox"},"confidence":{"type":"number","title":"Confidence"},"uncertainty":{"type":"number","title":"Uncertainty"},"L_mm":{"type":"number","title":"L Mm"},"W_mm":{"type":"number","title":"W Mm"},"area_mm2":{"type":"number","title":"Area Mm2"},"perimeter_mm":{"type":"number","title":"Perimeter Mm"},"aspect_ratio":{"type":"number","title":"Aspect Ratio"},"position":{"items":{"type":"number"},"type":"array","title":"Position"}},"type":"object","required":["id","class_id","class_name","bbox","confidence","uncertainty","L_mm","W_mm","area_mm2","perimeter_mm","aspect_ratio","position"],"title":"DefectOut"},"DetectResponse":{"properties":{"defects":{"items":{"$ref":"#/components/schemas/DefectOut"},"type":"array","title":"Defects"},"annotated_image":{"type":"string","title":"Annotated Image"}},"type":"object","required":["defects","annotated_image"],"title":"DetectResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IqiOut":{"properties":{"iqi_type":{"type":"string","title":"Iqi Type"},"achieved":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Achieved"},"required":{"type":"string","title":"Required"},"passed":{"type":"boolean","title":"Passed"}},"type":"object","required":["iqi_type","achieved","required","passed"],"title":"IqiOut"},"JudgeDefectIn":{"properties":{"id":{"type":"string","title":"Id"},"class_id":{"type":"integer","title":"Class Id"},"bbox":{"items":{"type":"number"},"type":"array","title":"Bbox"},"confidence":{"type":"number","title":"Confidence","default":0.5},"uncertainty":{"type":"number","title":"Uncertainty","default":0.5}},"type":"object","required":["id","class_id","bbox"],"title":"JudgeDefectIn"},"JudgeRequest":{"properties":{"base_metal_thickness_mm":{"type":"number","title":"Base Metal Thickness Mm"},"pixel_spacing_mm":{"type":"number","title":"Pixel Spacing Mm","default":1.0},"standard_id":{"type":"string","title":"Standard Id","default":"NB/T47013.2-2015"},"defects":{"items":{"$ref":"#/components/schemas/JudgeDefectIn"},"type":"array","title":"Defects","default":[]}},"type":"object","required":["base_metal_thickness_mm"],"title":"JudgeRequest"},"JudgeResponse":{"properties":{"joint_level":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Joint Level"},"per_defect_grade":{"items":{"type":"string"},"type":"array","title":"Per Defect Grade"},"basis":{"items":{"type":"string"},"type":"array","title":"Basis"},"need_review":{"type":"boolean","title":"Need Review"},"standard_id":{"type":"string","title":"Standard Id"},"standard_version":{"type":"string","title":"Standard Version"}},"type":"object","required":["joint_level","per_defect_grade","basis","need_review","standard_id","standard_version"],"title":"JudgeResponse"},"PreprocessResponse":{"properties":{"image_id":{"type":"string","title":"Image Id"},"thumbnail":{"type":"string","title":"Thumbnail"},"metrics":{"additionalProperties":{"type":"number"},"type":"object","title":"Metrics"}},"type":"object","required":["image_id","thumbnail","metrics"],"title":"PreprocessResponse"},"RecordsResponse":{"properties":{"items":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"},"stats":{"additionalProperties":true,"type":"object","title":"Stats"}},"type":"object","required":["items","total","stats"],"title":"RecordsResponse"},"ReportOut":{"properties":{"report_id":{"type":"string","title":"Report Id"},"image_id":{"type":"string","title":"Image Id"},"joint_level":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Joint Level"},"need_review":{"type":"boolean","title":"Need Review"},"evaluable":{"type":"boolean","title":"Evaluable"},"defect_count":{"type":"integer","title":"Defect Count"},"pdf_url":{"type":"string","title":"Pdf Url"}},"type":"object","required":["report_id","image_id","joint_level","need_review","evaluable","defect_count","pdf_url"],"title":"ReportOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VerifyResponse":{"properties":{"iqi":{"$ref":"#/components/schemas/IqiOut"},"density":{"type":"number","title":"Density"},"density_ok":{"type":"boolean","title":"Density Ok"},"evaluable":{"type":"boolean","title":"Evaluable"}},"type":"object","required":["iqi","density","density_ok","evaluable"],"title":"VerifyResponse"}}}}